+++ src/Exception/CannotAccessObjectException.php
